Fix Hd (English) bug on Samsung devices (5/6)
For some reason, several devices experience an oddity where the PRODUCT_LOCALE
being set to hdpi causes the language to be set to a made up setting called
"Hd (English)", and this causes usage of the market to crash repeatedly by trying
to display a non-existent currency. This only appears in devices with PRODUCT_LOCALE := hdpi
and I see most devices don't actually have this line, and removing it and replacing it
with the proper overrides seems to alleviate the problem.

Change-Id: Ic0640c09759c0bc234a975d3ad3f6c27192cbbdd